PointSample declaration is a bit more intelligent, can deal with the automatic observation cleaning that takes place in PointSource.
obs_clean is now in BaseSource, and will no longer fail if all data is bad.

Close issue #234
